refactor: 컴포넌트 변경된 구조에 맞게 atomic + pages 기준으로 변경
