+++ src/common/components/layout/index.tsx
@@ -3,7 +3,7 @@ import { css } from "@emotion/react";
 import { HTMLAttributes, ReactNode } from "react";
 import { Link } from "react-router-dom";
 import colors from "styles/color";
-import Menus from "common/components/menus";
+import Menus from "common/components/layout/menus";
 import { Search } from "common/components/icons";
 
 interface Props extends HTMLAttributes<HTMLDivElement> {
