+++ src/pages/menuDetail/components/index.ts
@@ -1,1 +1,3 @@
 export { default as MenuOption } from "./menuOption";
+export { default as LabelInput } from "./labelInput";
+export { default as LabelTextarea } from "./labelTextarea";
