+++ src/pages/menuDetail/components/menuOption/index.types.ts
@@ -1,0 +1,6 @@
+import { HTMLAttributes } from "react";
+
+export interface MenuProps extends HTMLAttributes<HTMLDivElement> {
+  title: string;
+  options: { name: string; price: number }[];
+}
